Split TestSCMDeleteBlocksCommandStatusManager and TestSCMDeleteBlocksCommandStatusManager
