+++ hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/node/SCMNodeManager.java
@@ -522,7 +522,7 @@ public class SCMNodeManager implements NodeManager {
         if (command.getType() == SCMCommandProto.Type.deleteBlocksCommand) {
           DeletedBlockLog deletedBlockLog = scmContext.getScm().
               getScmBlockManager().getDeletedBlockLog();
-          deletedBlockLog.getScmCommandStatusManager()
+          deletedBlockLog.getSCMDeletedBlockTransactionStatusManager()
               .onSent(datanodeDetails.getUuid(), command.getId());
         }
       }
